shuffled files to new paths, and a few related tweaks - no new functionality but made a basic plugin structure
