+++ connectors/php/plugins/rsc/cloudfiles_exceptions.php
