+++ connectors/php/plugins/rsc/share/magic.mgc
