+++ connectors/php/plugins/rsc/share/magic.mime
